trello_clone_django/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.10
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
sqlparse==0.5.5
typing_extensions==4.15.0
